backend/requirements.txt
fastapi==0.104.1
websockets>=10.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
